refactor(evaluation): add enums in the main seq evaluator
