+++ embeddings/evaluator/sequence_labeling_evaluator.py
@@ -1,4 +1,5 @@
-from typing import Any, Dict, Final, List, Literal, Optional, Sequence, Set, Union
+from enum import Enum
+from typing import Any, Dict, Final, List, Optional, Sequence, Set, Union
 
 import torch
 from numpy import typing as nptyping
@@ -10,20 +11,33 @@ from embeddings.metric.unit_seqeval_metric import UnitSeqevalMetric
 
 
 class SequenceLabelingEvaluator(MetricsEvaluator):
-    SEQEVAL_EVALUATION_MODES: Final[Set[str]] = {"conll", "strict"}
+    class EvaluationMode(str, Enum):
+        UNIT = "unit"
+        CONLL = "conll"
+        STRICT = "strict"
+
+    class TaggingScheme(str, Enum):
+        IOB1 = "IOB1"
+        IOB2 = "IOB2"
+        IOE1 = "IOE1"
+        IOE2 = "IOE2"
+        IOBES = "IOBES"
+        BILOU = "BILOU"
+
+    SEQEVAL_EVALUATION_MODES: Final[Set[str]] = {EvaluationMode.CONLL, EvaluationMode.STRICT}
 
     def __init__(
         self,
-        evaluation_mode: Literal["conll", "strict", "unit"] = "conll",
-        tagging_scheme: Optional[Literal["IOB1", "IOB2", "IOE1", "IOE2", "IOBES", "BILOU"]] = None,
+        evaluation_mode: EvaluationMode = EvaluationMode.CONLL,
+        tagging_scheme: Optional[TaggingScheme] = None,
     ) -> None:
         super().__init__()
         self.metric = self._get_metric(evaluation_mode, tagging_scheme)
 
     def _get_metric(
         self,
-        evaluation_mode: Literal["conll", "strict", "unit"],
-        tagging_scheme: Optional[Literal["IOB1", "IOB2", "IOE1", "IOE2", "IOBES", "BILOU"]] = None,
+        evaluation_mode: EvaluationMode,
+        tagging_scheme: Optional[TaggingScheme] = None,
     ) -> Union[HuggingFaceMetric, UnitSeqevalMetric]:
         if evaluation_mode in SequenceLabelingEvaluator.SEQEVAL_EVALUATION_MODES:
             if evaluation_mode == "strict" and not tagging_scheme:
