+++ examples/evaluate_sequence_labelling.py
@@ -1,29 +1,14 @@
 import pprint
-from enum import Enum
 from pathlib import Path
 from typing import Optional
 
 import typer
 
 from embeddings.defaults import RESULTS_PATH
+from embeddings.evaluator.sequence_labeling_evaluator import SequenceLabelingEvaluator
 from embeddings.pipeline.flair_sequence_labeling import FlairSequenceLabelingPipeline
 
 app = typer.Typer()
-
-
-class EvaluationMode(str, Enum):
-    UNIT = "unit"
-    CONLL = "conll"
-    STRICT = "strict"
-
-
-class TaggingScheme(str, Enum):
-    IOB1 = "IOB1"
-    IOB2 = "IOB2"
-    IOE1 = "IOE1"
-    IOE2 = "IOE2"
-    IOBES = "IOBES"
-    BILOU = "BILOU"
 
 
 def run(
@@ -41,16 +26,14 @@ def run(
     ),
     root: str = typer.Option(RESULTS_PATH.joinpath("pos_tagging")),
     hidden_size: int = typer.Option(256, help="Number of hidden states in RNN."),
-    evaluation_mode: EvaluationMode = typer.Option(
-        EvaluationMode.CONLL, help="Evaluation mode. Supported modes: [unit, conll, strict]."
+    evaluation_mode: SequenceLabelingEvaluator.EvaluationMode = typer.Option(
+        SequenceLabelingEvaluator.EvaluationMode.CONLL,
+        help="Evaluation mode. Supported modes: [unit, conll, strict].",
     ),
-    tagging_scheme: Optional[TaggingScheme] = typer.Option(
+    tagging_scheme: Optional[SequenceLabelingEvaluator.TaggingScheme] = typer.Option(
         None, help="Tagging scheme. Supported schemes: [IOB1, IOB2, IOE1, IOE2, IOBES, BILOU]"
     ),
 ) -> None:
-    if tagging_scheme:
-        tagging_scheme = tagging_scheme.value
-    evaluation_mode = evaluation_mode.value
     typer.echo(pprint.pformat(locals()))
 
     output_path = Path(root, embedding_name, dataset_name)
